v0.3
fixed bug with multiple files and column (hard work!).
- new option -c=,--cmt= to select the comment character in file [default =#]
- new option -lf=,--legend-file= to select the legend file [default=trans.info]
- syntax is better specified at the help prompt.
the general syntax for user defined operation is:
with the part in parentheses optional.
If file_index is not specified, the rule is intended to be applied to all files. This required to
modify the way @trans is treated in the script.
>> WARNING <<
if hybrid specifications are provided (e.g. #1 #2:2 #3) the one with explicit file_index (#2:2) is
postponed at the end of the list (--> #1 #3 #2:3) and the implicit ones are expanded for all files.
see legend file.
- if -l=,--line= is used results are returned in vertical form (one line per file), if -b,-wl,--with-line
flag is also used an additional counter is placed at first column.